% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTDA.ni.R
\docType{package}
\name{DTDA.ni-package}
\alias{DTDA.ni}
\alias{DTDA.ni-package}
\title{Package ‘DTDA.ni’}
\value{
\itemize{
\item{DTDAni: Implements a non-iterative estimator for the cumulative distribution of a doubly truncated variable}
\item{plot.DTDAni: S3 method to plot a DTDAni object by using the generic plot function.}
}
}
\description{
Non-iterative estimator for the cumulative distribution of a doubly truncated variable, see de Uña-Álvarez (2018).
Restricted to interval sampling.
}
\details{
Documentation for package ‘DTDA.ni’ version 1.0


\itemize{
\item{Package: DTDA.ni}
\item{Version: 1.0}
\item{Maintainer: José Carlos Soage González \email{jsoage@uvigo.es}}
\item{License: GPL-2}
}
}
\section{Acknowledgements}{

\itemize{
\item{Jacobo de Uña-Álvarez was supported by Grant MTM2014-55966-P, Spanish Ministry of Economy and Competitiveness.}
\item{José Carlos Soage was supported by Red Tecnológica de Matemática Industrial (Red TMATI), Cons. de Cultura, Educación e OU, Xunta de Galicia (ED341D R2016/051) and by Grupos de Referencia Competitiva, Consolidación y Estructuración de Unidades de Investigación Competitivas del SUG, Cons. de Cultura, Educación e OU, Xunta de Galicia (GRC ED431C 2016/040).}
}
}

\references{
de Uña-Álvarez J. (2018) A Non-iterative Estimator for Interval Sampling and
Doubly Truncated Data. In: Gil E., Gil E., Gil J., Gil M. (eds)
The Mathematics of the Uncertain. Studies in Systems, Decision and Control,
vol 142. Springer, Cham, pp. 387-400.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sidoruvigo/DTDA.ni}
}

}
\author{
\itemize{
 \item{de Uña-Álvarez, Jacobo.}
 \item{Soage González, José Carlos.}
}
}
